still a small bug

+++ run.py
@@ -65,7 +65,11 @@ def main(inputfile,
         err = container.logs(stdout=False, stderr=True)
         err_str = err.decode('utf-8')
         if not 'ERROR' in err_str:
-            shutil.copy(tmpOutputFolder+'/output.dat', output)
+            if not os.path.exists(tmpOutputFolder+'/output.dat'):
+                print('ERROR: Cannot find the output file: '+str(tmpOutputFolder+'/output.dat'))
+                print(err_str)
+            else:
+                shutil.copy(tmpOutputFolder+'/output.dat', output)
         else:
             print(err_str)
         container.remove()
